+++ kuryr_kubernetes/tests/unit/controller/handlers/test_vif.py
@@ -86,14 +86,12 @@ class TestVIFHandler(test_base.TestCase):
         self._activate_vif = self._handler._drv_vif_pool.activate_vif
         self._matc = self._handler._move_annotations_to_crd
         self._is_pod_scheduled = self._handler._is_pod_scheduled
-        self._is_pod_completed = self._handler._is_pod_completed
         self._request_additional_vifs = \
             self._multi_vif_drv.request_additional_vifs
 
         self._request_vif.return_value = self._vif
         self._request_additional_vifs.return_value = self._additioan_vifs
         self._is_pod_scheduled.return_value = True
-        self._is_pod_completed.return_value = False
         self._get_project.return_value = self._project_id
         self._get_subnets.return_value = self._subnets
         self._get_security_groups.return_value = self._security_groups
@@ -115,17 +113,6 @@ class TestVIFHandler(test_base.TestCase):
         self.assertFalse(h_vif.VIFHandler._is_pod_scheduled({'spec': {},
                                                              'status': {}}))
 
-    def test_is_pod_completed_pending(self):
-        self.assertFalse(h_vif.VIFHandler._is_pod_completed(self._pod))
-
-    def test_is_pod_completed_succeeded(self):
-        self.assertTrue(h_vif.VIFHandler._is_pod_completed({'status': {'phase':
-                        k_const.K8S_POD_STATUS_SUCCEEDED}}))
-
-    def test_is_pod_completed_failed(self):
-        self.assertTrue(h_vif.VIFHandler._is_pod_completed({'status': {'phase':
-                        k_const.K8S_POD_STATUS_FAILED}}))
-
     @mock.patch('kuryr_kubernetes.clients.get_kubernetes_client')
     @mock.patch('kuryr_kubernetes.controller.drivers.utils.is_host_network')
     @mock.patch('kuryr_kubernetes.controller.drivers.utils.get_kuryrport')
@@ -146,15 +133,17 @@ class TestVIFHandler(test_base.TestCase):
         self._request_additional_vifs.assert_not_called()
         self._activate_vif.assert_not_called()
 
+    @mock.patch('kuryr_kubernetes.utils.is_pod_completed')
     @mock.patch('kuryr_kubernetes.clients.get_kubernetes_client')
     @mock.patch('kuryr_kubernetes.controller.drivers.utils.is_host_network')
     @mock.patch('kuryr_kubernetes.controller.drivers.utils.get_kuryrport')
     def test_on_present_not_scheduled(self, m_get_kuryrport, m_host_network,
-                                      m_get_k8s_client):
+                                      m_get_k8s_client, m_is_pod_completed):
         m_get_kuryrport.return_value = self._kp
         m_host_network.return_value = False
         self._is_pod_scheduled.return_value = False
         self._matc.return_value = False
+        m_is_pod_completed.return_value = False
         k8s = mock.MagicMock()
         m_get_k8s_client.return_value = k8s
 
@@ -167,11 +156,13 @@ class TestVIFHandler(test_base.TestCase):
         self._request_additional_vifs.assert_not_called()
         self._activate_vif.assert_not_called()
 
+    @mock.patch('kuryr_kubernetes.utils.is_pod_completed')
     @mock.patch('kuryr_kubernetes.clients.get_kubernetes_client')
     @mock.patch('kuryr_kubernetes.controller.drivers.utils.get_kuryrport')
-    def test_on_present_on_completed_with_annotation(self, m_get_kuryrport,
-                                                     m_get_k8s_client):
-        self._is_pod_completed.return_value = True
+    def test_on_present_on_completed_with_kuryrport(self, m_get_kuryrport,
+                                                    m_get_k8s_client,
+                                                    m_is_pod_completed):
+        m_is_pod_completed.return_value = True
         m_get_kuryrport.return_value = self._kp
         self._matc.return_value = False
         k8s = mock.MagicMock()
@@ -179,30 +170,25 @@ class TestVIFHandler(test_base.TestCase):
 
         h_vif.VIFHandler.on_present(self._handler, self._pod)
 
-        k8s.add_finalizer.assert_called_once_with(self._pod,
-                                                  k_const.POD_FINALIZER)
-        self._matc.assert_called_once_with(self._pod)
         self._handler.on_finalize.assert_called_once_with(self._pod)
         self._request_vif.assert_not_called()
         self._request_additional_vifs.assert_not_called()
         self._activate_vif.assert_not_called()
 
+    @mock.patch('kuryr_kubernetes.utils.is_pod_completed')
     @mock.patch('kuryr_kubernetes.clients.get_kubernetes_client')
     @mock.patch('kuryr_kubernetes.controller.drivers.utils.get_kuryrport')
-    def test_on_present_on_completed_without_annotation(self, m_get_kuryrport,
-                                                        m_get_k8s_client):
-        self._is_pod_completed.return_value = True
+    def test_on_present_on_completed_without_kuryrport(self, m_get_kuryrport,
+                                                       m_get_k8s_client,
+                                                       m_is_pod_completed):
+        m_is_pod_completed.return_value = True
         m_get_kuryrport.return_value = None
-        self._matc.return_value = False
         k8s = mock.MagicMock()
         m_get_k8s_client.return_value = k8s
 
         h_vif.VIFHandler.on_present(self._handler, self._pod)
 
-        k8s.add_finalizer.assert_called_once_with(self._pod,
-                                                  k_const.POD_FINALIZER)
-        self._matc.assert_called_once_with(self._pod)
-        self._handler.on_finalize.assert_not_called()
+        self._handler.on_finalize.assert_called()
         self._request_vif.assert_not_called()
         self._request_additional_vifs.assert_not_called()
         self._activate_vif.assert_not_called()
